requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
Flask-Mail==0.9.1
python-dotenv==0.19.0
WTForms==3.0.0
email-validator==1.3.1